# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cybernix Nexus API Specification",
    "version": "2.0.0",
    "description": "Official OpenAPI 3.1 specification for Cybernix Nexus — NSEC Multi-Platform Competitive Programming Hub. Provides endpoints for student rating synchronization, real-time leaderboards, peer problem editorials, contest scheduling, and background cron scraping.",
    "contact": {
      "name": "Phoenix Tech Club (NSEC)",
      "url": "https://github.com/cybernixNexus"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    },
    {
      "url": "https://cybernixnexus.nsec.ac.in",
      "description": "Production Server"
    }
  ],
  "paths": {
...
    "/api/student/activity": {
      "get": {
        "summary": "Compute Student 52-Week Solve Heatmap & Streak",
        "description": "Aggregates historical solve activity from DailySnapshot records to render 52-week activity matrices and compute active consecutive day streaks.",
        "tags": [
          "Student Management"
        ],
        "parameters": [
          {
            "name": "studentId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Target student ID (defaults to active authenticated student)"
          }
        ],
        "responses": {
          "200": {
            "description": "Activity heatmap matrix and streak metrics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "currentStreak": {
                      "type": "integer",
                      "example": 14
                    },
                    "maxStreak": {
                      "type": "integer",
                      "example": 30
                    },
                    "totalAnnualSolves": {
                      "type": "integer",
                      "example": 420
                    },
                    "gridData": {
                      "type": "array",
                      "items": {
                        "type": "array",
                        "items": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 4
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/student": {
      "get": {
        "summary": "Get Current Student Profile",
        "description": "Returns the full student profile, linked platform handles, solve stats, college rank, and sync job status.",
        "tags": [
          "Student Management"
        ],
        "responses": {
          "200": {
            "description": "Student profile details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "student": {
                      "$ref": "#/components/schemas/StudentDetail"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "image": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create / Update Student Profile",
        "description": "Updates student department, graduation year, and platform handles for authenticated user.",
        "tags": [
          "Student Management"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student profile updated successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token for background cron operations (CRON_SECRET)."
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth persistent session cookie."
      }
    },
    "schemas": {
...
      "UpdateStudentRequest": {
        "type": "object",
        "properties": {
          "department": {
            "type": "string",
            "example": "CSE"
          },
          "graduationYear": {
            "type": "integer",
            "example": 2026
          },
          "rollNumber": {
            "type": "string"
          },
          "github": {
            "type": "string"
          },
          "linkedin": {
            "type": "string"
          },
          "leetcode": {
            "type": "string"
          },
          "codeforces": {
            "type": "string"
          },
          "gfg": {
            "type": "string"
          },
          "codechef": {
            "type": "string"
          }
        }
      },
      "StudentStats": {
        "type": "object",
        "properties": {
          "leetcodeSolved": {
            "type": "integer",
            "example": 450
          },
          "leetcodeRating": {
            "type": "integer",
            "example": 1980
          },
          "codeforcesRating": {
            "type": "integer",
            "example": 1740
          },
          "codeforcesMaxRating": {
            "type": "integer",
            "example": 1820
          },
          "codeforcesRank": {
            "type": "string",
            "example": "expert"
          },
          "codeforcesSolved": {
            "type": "integer",
            "example": 280
          },
          "gfgScore": {
            "type": "integer",
            "example": 380
          },
          "codechefRating": {
            "type": "integer",
            "example": 1790
          },
          "totalScore": {
            "type": "integer",
            "example": 32400
          },
          "ranking": {
            "type": "integer",
            "example": 3
          },
          "departmentRanking": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "StudentDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "rollNumber": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "graduationYear": {
            "type": "integer"
          },
          "github": {
            "type": "string",
            "nullable": true
          },
          "linkedin": {
            "type": "string",
            "nullable": true
          },
          "leetcode": {
            "type": "string",
            "nullable": true
          },
          "codeforces": {
            "type": "string",
            "nullable": true
          },
          "gfg": {
            "type": "string",
            "nullable": true
          },
          "codechef": {
            "type": "string",
            "nullable": true
          },
          "stats": {
            "$ref": "#/components/schemas/StudentStats"
          }
        }
      },
...
    }
  }
}